requirements.txt
asgiref==3.6.0
autopep8==2.0.2
crispy-bootstrap5==0.7
Django==4.1.7
django-choices==1.7.2
django-crispy-forms==2.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
